src/analysis/solver/optimization.ts
import { SolverState } from '../../../spec/Solver';
import { Matrix, createMatrix, multiplyMatrix } from './matrix';

export function optimizeSolution(state: SolverState): SolverState {
  // Check if current solution can be improved
  const currentValue = evaluateObjective(state);
  const improvedSolution = findImprovedSolution(state, currentValue);
  
  if (improvedSolution) {
    return {
      ...state,
      problem: {
        ...state.problem,
        variables: improvedSolution.variables
      }
    };
  }
  
  return state;
}

function evaluateObjective(state: SolverState): number {
  // Evaluate the objective function at current point
  // This is a placeholder implementation
  return 0;
}

function findImprovedSolution(state: SolverState, currentValue: number): {
  variables: typeof state.problem.variables;
  value: number;
} | null {
  // Try to find an improved solution
  // This is a placeholder implementation
  return null;
}

export function checkFeasibility(state: SolverState): boolean {
  // Check if all constraints are satisfied
  for (const constraint of state.problem.constraints) {
    if (!isConstraintSatisfied(state, constraint)) {
      return false;
    }
  }
  return true;
}

function isConstraintSatisfied(
  state: SolverState,
  constraint: typeof state.problem.constraints[0]
): boolean {
  // Check if a single constraint is satisfied
  // This is a placeholder implementation
  return true;
}

export function findInitialSolution(state: SolverState): SolverState {
  // Find a basic feasible solution
  // This is a placeholder implementation
  return state;
}
